requirements.txt
starlette
aiofiles
uvicorn
eventlet
python-socketio
aiohttp
opencv-python
torch
torchvision
pandas
numpy
imutils
Pillow
python-dotenv